requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
et-xmlfile==1.1.0
htmlmin==0.1.12
idna==3.4
lxml==4.9.3
openpyxl==3.1.2
python-docx==1.0.1
requests==2.31.0
soupsieve==2.5
typing_extensions==4.8.0
urllib3==2.0.7
